fixed bug and added foundation functions
fixed small error in which a space would be added before instructions after labels. Made binary to int and int to binary converter for instructions that need you to take in 1 bit or halfword etc. Created a function that fully splits the entire line into words so you can easily create functions without thinking to hard about the backend of how the registers are split, created setter function for registers.